+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/composable/ComposableFapButtons.kt
@@ -14,6 +14,7 @@ import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.faphub.installation.button.api.FapButtonSize
 import com.flipperdevices.faphub.installation.button.impl.R
 import com.flipperdevices.faphub.installation.button.impl.composable.dialogs.ComposableFlipperNotConnectedDialog
+import com.flipperdevices.faphub.installation.button.impl.composable.dialogs.ComposableFlipperNotSdCardDialog
 import com.flipperdevices.faphub.installation.button.impl.composable.elements.ComposableFlipperButton
 import com.flipperdevices.faphub.installation.button.impl.composable.elements.ComposableInProgressFapButton
 
@@ -33,16 +34,36 @@ fun ComposableFapInstallButton(
 }
 
 @Composable
+fun ComposableFlipperNoSdCardButton(
+    fapButtonSize: FapButtonSize,
+    modifier: Modifier = Modifier
+) {
+    var showDialog by remember { mutableStateOf(false) }
+    ComposableFapInstallButton(
+        modifier = modifier,
+        fapButtonSize = fapButtonSize,
+        onClick = {
+            showDialog = true
+        }
+    )
+    if (showDialog) {
+        ComposableFlipperNotSdCardDialog(
+            onDismiss = {
+                showDialog = false
+            },
+        )
+    }
+}
+
+@Composable
 fun ComposableFlipperNotConnectedButton(
     fapButtonSize: FapButtonSize,
     onOpenDeviceTab: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     var showDialog by remember { mutableStateOf(false) }
-    ComposableFlipperButton(
+    ComposableFapInstallButton(
         modifier = modifier,
-        text = stringResource(R.string.faphub_installation_install),
-        color = LocalPallet.current.accent,
         fapButtonSize = fapButtonSize,
         onClick = {
             showDialog = true