+++ components/faphub/installation/stateprovider/api/src/main/java/com/flipperdevices/faphub/installation/stateprovider/api/model/FapState.kt
@@ -11,7 +11,6 @@ sealed class FapState {
     data object Installed : FapState()
 
     data object ReadyToInstall : FapState()
-    data object ConnectFlipper : FapState()
 
     data class NotAvailableForInstall(val reason: NotAvailableReason) : FapState()
 
@@ -38,5 +37,7 @@ enum class NotAvailableReason {
     BUILD_RUNNING,
     UNSUPPORTED_APP,
     FLIPPER_OUTDATED,
-    UNSUPPORTED_SDK
+    UNSUPPORTED_SDK,
+    NO_SD_CARD,
+    FLIPPER_NOT_CONNECTED
 }