+++ components/faphub/installedtab/impl/src/main/java/com/flipperdevices/faphub/installedtab/impl/model/FapBatchUpdateButtonState.kt
@@ -1,12 +1,12 @@
 package com.flipperdevices.faphub.installedtab.impl.model
 
 sealed class FapBatchUpdateButtonState {
-    object Loading : FapBatchUpdateButtonState()
+    data object Loading : FapBatchUpdateButtonState()
 
     data class ReadyToUpdate(val count: Int) : FapBatchUpdateButtonState()
 
-    object Offline : FapBatchUpdateButtonState()
-    object NoUpdates : FapBatchUpdateButtonState()
+    data object Offline : FapBatchUpdateButtonState()
+    data object NoUpdates : FapBatchUpdateButtonState()
 
-    object UpdatingInProgress : FapBatchUpdateButtonState()
+    data object UpdatingInProgress : FapBatchUpdateButtonState()
 }