+++ components/faphub/installedtab/impl/src/main/java/com/flipperdevices/faphub/installedtab/impl/viewmodel/InstalledFapsViewModel.kt
@@ -149,7 +149,7 @@ class InstalledFapsViewModel @VMInject constructor(
 }
 
 sealed class FapInstalledInternalLoadingState {
-    object Loading : FapInstalledInternalLoadingState()
+    data object Loading : FapInstalledInternalLoadingState()
 
     data class LoadedOffline(
         val faps: ImmutableList<OfflineFapApp>