+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/dialogs/ComposableDialogPending.kt
@@ -10,7 +10,6 @@ import com.flipperdevices.core.ui.dialog.composable.multichoice.FlipperMultiChoi
 import com.flipperdevices.core.ui.dialog.composable.multichoice.FlipperMultiChoiceDialogModel
 import com.flipperdevices.updater.card.R
 import com.flipperdevices.updater.model.FirmwareVersion
-import com.flipperdevices.core.ui.res.R as DesignSystem
 
 @Composable
 fun FlipperDialogReadyUpdate(
@@ -93,14 +92,8 @@ fun FlipperDialogSynchronization(
 
 @Composable
 fun FlipperDialogFileVeryBig(onDismiss: () -> Unit) {
-    val imageId = if (MaterialTheme.colors.isLight) {
-        DesignSystem.drawable.ic_firmware_application_deprecated
-    } else {
-        DesignSystem.drawable.ic_firmware_application_deprecated_dark
-    }
-
     FlipperDialog(
-        imageId = imageId,
+        imageId = R.drawable.pic_wrong_file,
         titleId = R.string.update_card_dialog_file_big_title,
         textId = R.string.update_card_dialog_file_big_desc,
         buttonTextId = R.string.update_card_dialog_battery_btn,
@@ -111,14 +104,8 @@ fun FlipperDialogFileVeryBig(onDismiss: () -> Unit) {
 
 @Composable
 fun FlipperDialogFileExtension(onDismiss: () -> Unit) {
-    val imageId = if (MaterialTheme.colors.isLight) {
-        DesignSystem.drawable.ic_firmware_application_deprecated
-    } else {
-        DesignSystem.drawable.ic_firmware_application_deprecated_dark
-    }
-
     FlipperDialog(
-        imageId = imageId,
+        imageId = R.drawable.pic_wrong_file,
         titleId = R.string.update_card_dialog_extension_title,
         textId = R.string.update_card_dialog_extension_desc,
         buttonTextId = R.string.update_card_dialog_battery_btn,